pyproject.toml
[tool.poetry]
name = "drfpro"
version = "0.1.0"
description = ""
authors = ["flizarrq <kirillozubenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
mysqlclient = "^2.2.4"
django-filter = "^24.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
django-dotenv = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
